# MLA: allow Q8_0 K-cache for MLA